实现回复以及展示一级评论功能

+++ src/main/java/com/wqy/campusbbs/controller/PublishController.java
@@ -1,7 +1,6 @@
 package com.wqy.campusbbs.controller;
 
 import com.wqy.campusbbs.dto.QuestionDTO;
-import com.wqy.campusbbs.mapper.QuestionMapper;
 import com.wqy.campusbbs.model.Question;
 import com.wqy.campusbbs.model.User;
 import com.wqy.campusbbs.service.QuestionService;
@@ -21,7 +20,7 @@ public class PublishController {
     private QuestionService questionService;
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id,
+    public String edit(@PathVariable(name = "id") Long id,
                        Model model) {
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title", question.getTitle());
@@ -40,7 +39,7 @@ public class PublishController {
     public String doPublish(@RequestParam(value = "title", required = false) String title,
                             @RequestParam(value = "description", required = false) String description,
                             @RequestParam(value = "tag", required = false) String tag,
-                            @RequestParam(value = "id", required = false) Integer id,
+                            @RequestParam(value = "id", required = false) Long id,
                             HttpServletRequest request,
                             Model model) {
         model.addAttribute("title", title);
